// src/app/models/media-object.model.ts
export interface MediaObject {
  _id?: string;
  name?: string;
  alternativeText?: string;
  caption?: string;
  width?: number;
  height?: number;
  hash?: string;
  ext?: string;
  mime?: string;
  size?: number;
  url?: string;
  path?: string;
  provider?: string;
  provider_metadata?: string;
  thumbnail?: Thumbnail;
  createdAt?: string;
  updatedAt?: string;
}

export interface Thumbnail {
  width: number;
  height: number;
  hash: string;
  ext: string;
  mime: string;
  size: number;
  url: string;
  path: string;
}